requirements.txt
google_api_python_client==2.46.0
google_auth_oauthlib==0.5.1
itemadapter==0.5.0
pandas==1.4.2
protobuf==3.20.1
Scrapy==2.6.1
